requirements.txt
appdirs==1.4.4
PySimpleGUI==4.38.0
PyYAML==5.4.1